src/main/java/com/scoot/api/exceptionHandler/GlobalExceptionHandler.java
package com.scoot.api.exceptionHandler;

import com.scoot.api.exceptionHandler.exception.InvalidRequestException;
import com.scoot.api.exceptionHandler.exception.RateLimitExceededException;
import com.scoot.api.exceptionHandler.exception.UnknownApiKeyException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class GlobalExceptionHandler {

    @ExceptionHandler(InvalidRequestException.class)
    public ResponseEntity<String> handleInvalidRequestException(InvalidRequestException ex) {
        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ex.getMessage());
    }

    @ExceptionHandler(UnknownApiKeyException.class)
    public ResponseEntity<String> handleUnknownApiKeyException(UnknownApiKeyException ex) {
        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ex.getMessage());
    }

    @ExceptionHandler(RateLimitExceededException.class)
    public ResponseEntity<String> handleRateLimitExceededException(RateLimitExceededException ex) {
        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body(ex.getMessage());
    }
}

